#556 feat(scheduler): Add the next schedule calculation to DeviceSingleSchedule
