+++ services/scheduler/scheduler/services/device_single_schedule.py
@@ -1,6 +1,7 @@
-from datetime import datetime
+from datetime import datetime, time
 from typing import Any, Dict
 
+import pytz
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 from apscheduler.triggers.date import DateTrigger
 from dependency_injector import providers
@@ -45,8 +46,10 @@ class DeviceSingleSchedule(DeviceSchedule):
         self.__at: str = device_schedule['at']
 
     def build_trigger(self, start: datetime | None = None):
+        at = self.__calculate_date(start)
+
         trigger = DateTrigger(
-            run_date=start
+            run_date=at
         )
 
         params = None
@@ -58,6 +61,22 @@ class DeviceSingleSchedule(DeviceSchedule):
 
     def _check_next_condition(self, **_):
         return True
+
+    def __calculate_date(self, start: datetime | None = None):
+        at = [int(part) for part in self.__at.split(':', 3)]
+
+        timezone = self._timezone
+
+        start_date = timezone.localize(datetime.combine(
+            start.date(), time(at[0], at[1], at[2], 0)
+        ))
+
+        # find the next appropriate day-of-week
+        start_date = self._find_valid_day(start_date)
+
+        start_date = start_date.astimezone(pytz.UTC)
+
+        return start_date
 
     def __str__(self):
         builder = f'Every {self.__at}'
